+++ inst/doc/intro.R
@@ -20,17 +20,16 @@ towns
 water
 
 ## ----layers, fig.align='center', fig.width=5, fig.height=5, fig.cap='Visualization of the \\texttt{water}, \\texttt{towns} and \\texttt{cities} layers'----
-plot(st_geometry(towns), col = NA)
-plot(st_geometry(water), col = "lightblue", add = TRUE)
+plot(st_geometry(water), col = "lightblue")
 plot(st_geometry(towns), col = "grey", pch = 1, add = TRUE)
 plot(st_geometry(cities), col = "red", pch = 1, add = TRUE)
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], progress = FALSE)
+nn = st_nn(cities, towns[1:5, ])
 nn
 
 ## ------------------------------------------------------------------------
-l = st_connect(cities, towns[1:5, ], ids = nn, progress = FALSE)
+l = st_connect(cities, towns[1:5, ], ids = nn)
 l
 
 ## ----st_connect, fig.align='center', fig.width=5, fig.height=5, fig.cap="Nearest neighbor match between \\texttt{cities} (in red) and \\texttt{towns[1:5, ]} (in grey)"----
@@ -41,23 +40,23 @@ text(st_coordinates(cities)[, 1], st_coordinates(cities)[, 2], 1:3, col = "red",
 text(st_coordinates(towns[1:5, ])[, 1], st_coordinates(towns[1:5, ])[, 2], 1:5, pos = 4)
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], sparse = FALSE, progress = FALSE)
+nn = st_nn(cities, towns[1:5, ], sparse = FALSE)
 nn
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], k = 2, progress = FALSE)
+nn = st_nn(cities, towns[1:5, ], k = 2)
 nn
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], sparse = FALSE, k = 2, progress = FALSE)
+nn = st_nn(cities, towns[1:5, ], sparse = FALSE, k = 2)
 nn
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], k = 2, returnDist = TRUE, progress = FALSE)
+nn = st_nn(cities, towns[1:5, ], k = 2, returnDist = TRUE)
 nn
 
 ## ------------------------------------------------------------------------
-nn = st_nn(cities, towns[1:5, ], k = 2, maxdist = 50000, progress = FALSE)
+nn = st_nn(cities, towns[1:5, ], k = 2, maxdist = 50000)
 nn
 
 ## ---- results='hide'-----------------------------------------------------
@@ -71,15 +70,15 @@ x = st_nn(cities, towns, k = 10)
 l = st_connect(cities, towns, ids = x)
 
 ## ----cities_towns, fig.align='center', fig.width=5, fig.height=5, warning=FALSE, fig.cap="Nearest 10 \\texttt{towns} features from each \\texttt{cities} feature"----
-plot(st_geometry(towns), col = "darkgrey")
-plot(st_geometry(l), add = TRUE)
+plot(st_geometry(l))
 plot(st_geometry(cities), col = "red", add = TRUE)
+plot(st_geometry(towns), col = "darkgrey", add = TRUE)
 
 ## ------------------------------------------------------------------------
-nn = st_nn(water[-1, ], towns, k = 20, progress = FALSE)
+nn = st_nn(water[-1, ], towns, k = 20)
 
 ## ---- warning=FALSE------------------------------------------------------
-l = st_connect(water[-1, ], towns, ids = nn, progress = FALSE, dist = 100)
+l = st_connect(water[-1, ], towns, ids = nn, dist = 100)
 
 ## ----water_towns, fig.align='center', fig.width=5, fig.height=5, warning=FALSE, fig.cap="Nearest 20 \\texttt{towns} features from each \\texttt{water} polygon"----
 plot(st_geometry(water[-1, ]), col = "lightblue", border = "grey")