+++ NEWS.md
@@ -55,13 +55,28 @@
 * Added 'raster_extract', a wrapper around 'raster::extract'
 * Added 'raster_trim', a wrapper around 'raster::trim'
 
-## nngeo 0.3.0
+## nngeo 0.3.0 (2019-12-03)
 
 * Added 'raster_extend', a wrapper around 'raster::extend'
 * When using 'returnDist=TRUE', distances are now returned as sparse 'list' rather than a 'matrix'
 
+## nngeo 0.3.4
+
+* 'st_nn' with 'returnDist=TRUE' returns named list with elements 'nn' and 'dist'
+* Replaced 'towns' dataset
+* Removed 'raster_*" functions
+* Added 'focal2' function (a 3x3 focal filter on 'stars')
+* Added 'line' and 'pnt' sample data (based on pgRouting tutorial)
+* Added 'st_split_junctions' function
+* 'st_connect' now uses 'st_nearest_point' rather than point sampling
+* 'st_connect' removes CRS before calculating nearest point, to omit the warning when using lon-lat
+
 ## Other ideas:
 
+* Add 'split line to equal parts' function
+* Add 'round coord' function
+* Add 'extend lines' function
+* Add 'bridge lines to network' function
 * Add parallel processing option
 * Add 'st_az' function
 * Geodesic buffer